[IOS] refactor: model 분리
- Issue.swift에 있는 모델 객체 개별 파일들로 분리
- IssueElement 삭제
