+++ ios/IssueTracker/IssueTracker/Models/Issue.swift
@@ -7,15 +7,6 @@
 
 import Foundation
 
-struct IssueElement: Codable {
-    let issues: [Issue]
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.singleValueContainer()
-        issues = try container.decode([Issue].self)
-    }
-}
-
 struct Issue: Codable {
     let title: String
     let issueID: Int
@@ -23,7 +14,7 @@ struct Issue: Codable {
     let isOpen: Int
     let userAuthor, user: User?
     let issueLabels: [IssueLabel]
-    let milestoneTitle: MileStone?
+    let milestoneTitle: Milestone?
     let comments: [Comment]
 
     enum CodingKeys: String, CodingKey {
@@ -38,39 +29,3 @@ struct Issue: Codable {
     }
 }
 
-struct Comment: Codable {
-    let commentID: Int
-    let userID: String
-
-    enum CodingKeys: String, CodingKey {
-        case commentID = "commentId"
-        case userID = "userId"
-    }
-}
-
-struct IssueLabel: Codable {
-    let labelID: Int
-    let label: Label
-
-    enum CodingKeys: String, CodingKey {
-        case labelID = "id"
-        case label = "Label"
-    }
-}
-
-struct Label: Codable {
-    let labelName, color: String
-}
-
-struct MileStone: Codable {
-    let title: String
-}
-
-struct User: Codable {
-    let userID, profileURL: String
-
-    enum CodingKeys: String, CodingKey {
-        case userID = "userId"
-        case profileURL = "profile_url"
-    }
-}
