+++ ios/IssueTracker/IssueTracker/Models/Milestone.swift
@@ -1,0 +1,12 @@
+//
+//  Milestone.swift
+//  IssueTracker
+//
+//  Created by 김석호 on 2020/11/09.
+//
+
+import Foundation
+
+struct Milestone: Codable {
+    let title: String
+}
