+++ ios/IssueTracker/IssueTracker/Models/User.swift
@@ -1,0 +1,17 @@
+//
+//  User.swift
+//  IssueTracker
+//
+//  Created by 김석호 on 2020/11/09.
+//
+
+import Foundation
+
+struct User: Codable {
+    let userID, profileURL: String
+
+    enum CodingKeys: String, CodingKey {
+        case userID = "userId"
+        case profileURL = "profile_url"
+    }
+}
